requirements.txt
mcp>=1.0.0
solders>=0.21.0
solana>=0.34.0
httpx>=0.25.0
asyncio-mqtt>=0.11.0
pydantic>=2.0.0
python-dotenv>=1.0.0
base58>=2.1.0